refactor: yml 주입방식 변경 (테스트 용이와 간편함을 위해) 및 불필요한 속성 Config 이전
