+++ backend/pcloud-api/src/main/java/com/api/config/scan/ComponentScanConfig.java
@@ -1,9 +1,9 @@
-package com.api.config;
+package com.api.config.scan;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-@ComponentScan(value = {"com.common", "com.domain"})
+@ComponentScan(value = {"com.common", "com.domain", "com.infrastructure"})
 public class ComponentScanConfig {
 }
