+++ backend/pcloud-api/src/main/resources/application.yml
@@ -1,9 +1,6 @@
 spring:
   profiles:
-    include:
-      - domain
-      - common
-      - infrastructure
+    active: local
 
 management:
   server:
