+++ backend/pcloud-api/src/test/java/com/api/PCloudApiServerTest.java
@@ -3,7 +3,7 @@ package com.api;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-//@SpringBootTest
+@SpringBootTest
 class PCloudApiServerTest {
 
     @Test
