+++ backend/pcloud-api/src/test/resources/application.yml
@@ -1,10 +1,7 @@
 spring:
   profiles:
     include:
-      - domain
-      - common
-      - infrastructure
-
-jasypt:
-  encryptor:
-    password: password
+      - domain-test
+      - common-test
+      - infrastructure-test
+    active: test
