+++ backend/pcloud-common/src/main/java/com/common/config/YamlPropertySourceFactory.java
@@ -1,0 +1,16 @@
+package com.common.config;
+
+import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
+import org.springframework.core.env.PropertiesPropertySource;
+import org.springframework.core.io.support.DefaultPropertySourceFactory;
+import org.springframework.core.io.support.EncodedResource;
+
+public class YamlPropertySourceFactory extends DefaultPropertySourceFactory {
+
+    @Override
+    public org.springframework.core.env.PropertySource<?> createPropertySource(final String name, final EncodedResource resource) {
+        YamlPropertiesFactoryBean factory = new YamlPropertiesFactoryBean();
+        factory.setResources(resource.getResource());
+        return new PropertiesPropertySource(resource.getResource().getFilename(), factory.getObject());
+    }
+}
