+++ backend/pcloud-common/src/main/resources/application-common-prod.yml
@@ -1,0 +1,8 @@
+jasypt:
+  encryptor:
+    bean: jasyptEncryptor
+    password: ${ENCRYPT_KEY}
+
+jwt:
+  secret: ENC(hQLPUuufo6+rX3SKuCA1WOTUqjyLREZsBF1q4dFrB2+WLMQNVyGhB4d2zskhI6sdCWgDaPBDABZFIg6nk3eLtZErUieH1I33)
+  expiration-period: ENC(akNgWoo0R4MHHtv0+2iBQw==)
