+++ backend/pcloud-common/src/test/java/com/common/TestConfig.java
@@ -1,0 +1,9 @@
+package com.common;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+
+@SpringBootApplication
+@ConfigurationPropertiesScan
+public class TestConfig {
+}
