+++ backend/pcloud-domain/src/main/resources/application-domain-prod.yml
@@ -1,0 +1,14 @@
+spring:
+  datasource:
+    url: jdbc:mysql://prod:3306/pcloud?useSSL=false&serverTimezone=UTC&useUnicode=true&characterEncoding=utf8&allowPublicKeyRetrieval=true
+    username: prod
+    password: prod
+    driver-class-name: com.mysql.cj.jdbc.Driver
+
+  flyway:
+    enabled: true
+    baseline-on-migrate: true
+
+  jpa:
+    hibernate:
+      ddl-auto: validate
