--- backend/pcloud-domain/src/main/resources/application-domain.yml
@@ -1,42 +1,0 @@
----
-spring:
-  config:
-    activate:
-      on-profile: local
-  datasource:
-    url: jdbc:mysql://localhost:3306/pcloud?useSSL=false&serverTimezone=UTC&useUnicode=true&characterEncoding=utf8&allowPublicKeyRetrieval=true
-    username: root
-    password: root
-    driver-class-name: com.mysql.cj.jdbc.Driver
-
-  flyway:
-    enabled: false
-    baseline-on-migrate: false
-
-  jpa:
-    hibernate:
-      ddl-auto: create
-
-    properties:
-      hibernate:
-        format_sql: true
-        show_sql: true
----
-spring:
-  config:
-    activate:
-      on-profile: prod
-  datasource:
-    url: jdbc:mysql://prod:3306/pcloud?useSSL=false&serverTimezone=UTC&useUnicode=true&characterEncoding=utf8&allowPublicKeyRetrieval=true
-    username: prod
-    password: prod
-    driver-class-name: com.mysql.cj.jdbc.Driver
-
-  flyway:
-    enabled: true
-    baseline-on-migrate: true
-
-  jpa:
-    hibernate:
-      ddl-auto: validate
----
